+++ setup.py
@@ -1,0 +1,7 @@
+from setuptools import setup, find_packages
+
+setup(
+    name = "Continuous-Integration-Group10",
+    version = "0.1.0",
+    packages = find_packages(),
+)
