--- tests/test.py
@@ -1,11 +1,0 @@
-import unittest
-
-class TestFunctions(unittest.TestCase):
-
-    # structure of how the tests should be written.
-    # write 'test_' followed by a name that describes what the test does.
-    def test_nameOfTest(self):
-        self.assertTrue(True)
-
-if __name__ == '__main__':
-    unittest.main()
